+++ R/PipeOpTargetTrafoBoxCox.R
@@ -56,6 +56,7 @@ PipeOpTargetTrafoBoxCox = R6Class(
     initialize = function(id = "fcst.targetboxcox", param_vals = list()) {
       param_set = ps(
         lambda = p_dbl(default = NULL, special_vals = list(NULL), tags = "train"),
+        period = p_uty(default = NULL, custom_check = check_period),
         method = p_fct(c("guerrero", "loglik"), default = "guerrero", tags = c("train", "estimate")),
         lower = p_dbl(default = -1, tags = c("train", "estimate")),
         upper = p_dbl(default = 2, tags = c("train", "estimate"))
@@ -78,12 +79,16 @@ PipeOpTargetTrafoBoxCox = R6Class(
       key_cols = task$col_roles$key
       if (!is.null(lambda) || length(key_cols) == 0L) {
         if (is.null(lambda)) {
-          lambda = invoke(forecast::BoxCox.lambda, as.ts(task), .args = self$param_set$get_values(tags = "estimate"))
+          lambda = invoke(
+            forecast::BoxCox.lambda,
+            as.ts(task, period = self$param_set$values$period),
+            .args = self$param_set$get_values(tags = "estimate")
+          )
         }
         return(list(lambda = lambda))
       }
       target = task$target_names
-      period = freq_to_period(task$freq)
+      period = task_period(self$param_set$values$period, task)
       args = self$param_set$get_values(tags = "estimate")
       estimate_lambda = function(y) {
         invoke(forecast::BoxCox.lambda, stats::ts(as.numeric(y), frequency = period), .args = args)